Replace e()/$e style error reporting with XH_message()
These messages are so much nicer.  We also do no longer report the file
names, but rather the names of the text blocks, what also solves
potential information leakage in the front-end.

Furthermore, we tuck `XH_message()` away in a private method, which
uses new style variadic arguments with the splat operator.  That
requires PHP ≥ 5.6.0, what shouldn't be a real problem in practice,
though.