Drop now superfluous attack-tests
We're testing the CSRF protection via our unit tests (well, only mocks,
but that should be sufficient), so there is no more need for the
attack-tests.

Since we're only left with the "unit-tests" which are not really unit-
tests, but rather sociable tests, so we get rid of the "unit".